rework of a consolidation script && add a DB user setup script
