+++ application/.env.example
@@ -11,7 +11,7 @@ LOG_CHANNEL=stack
 LOG_DEPRECATIONS_CHANNEL=null
 LOG_LEVEL=debug
 
-DB_CONNECTION=pgsql_app
+DB_CONNECTION=pgsql
 DB_DATABASE=app
 DB_HOST=postgres
 DB_PORT=5432
@@ -20,8 +20,6 @@ DB_PASSWORD=pass
 
 PG_APP_SCHEMA=application
 PG_SYNC_SCHEMA=entity_cache
-
-PG_SYNC_CONNECTION_NAME=pgsql_sync
 
 AMQP_HOST=rabbitmq
 # Should be the same as ../.env AMQP_PORT
